requirements.txt
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.13.3
markdownify~=1.1.0
flask~=3.1.0
discord~=2.3.2
notion-client~=2.3.0
openai~=1.66.3
firecrawl-py~=0.1.0
python-dotenv~=1.0.1